+++ src/components/TableListInPage.tsx
@@ -10,7 +10,7 @@ import { VirtualListFooter } from '@/components/VirtualList/VirtualListFooter.js
 import { VirtualTableList, type VirtualTableListProps } from '@/components/VirtualList/VirtualTableList.js';
 import { EMPTY_OBJECT } from '@/constants/index.js';
 import { classNames } from '@/helpers/classNames.js';
-import { narrowToSocialSource } from '@/helpers/narrowSource.js';
+import { narrowToSocialSource } from '@/helpers/narrowToSocialSource.js';
 import { useIsLogin } from '@/hooks/useIsLogin.js';
 import { useGlobalState } from '@/store/useGlobalStore.js';
